Entrega Final

+++ src/main/java/com/senai/alexandre/consultasmedicas/service/MedicoService.java
@@ -1,5 +1,6 @@
 package com.senai.alexandre.consultasmedicas.service;
 
+import com.senai.alexandre.consultasmedicas.exception.MedicoException;
 import com.senai.alexandre.consultasmedicas.model.Medico;
 import com.senai.alexandre.consultasmedicas.repository.MedicoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +15,7 @@ public class MedicoService {
     private MedicoRepository medicoRepository;
 
     public Medico save(Medico medico){
+        validarMedico(medico);
         return this.medicoRepository.save(medico);
     }
 
@@ -26,7 +28,25 @@ public class MedicoService {
     }
 
     public void delete(Integer id){
+        if (!medicoRepository.existsById(id)) {
+            throw new MedicoException("Paciente não encontrado para exclusão.");
+        }
         this.medicoRepository.deleteById(id);
+    }
+
+    // Tratamento de Exceções
+    private void validarMedico(Medico medico) {
+        if (medico.getNome() == null || medico.getNome().trim().isEmpty()) {
+            throw new MedicoException("O nome do medico é obrigatório.");
+        }
+
+        if (medico.getEspecialidade() == null || medico.getEspecialidade().trim().isEmpty()) {
+            throw new MedicoException("A especialidade do medico é obrigatória.");
+        }
+
+        if (medico.getCrm() == null || medico.getCrm().trim().isEmpty()) {
+            throw new MedicoException("O crm do medico é obrigatório.");
+        }
     }
 
 }
